Use abstract stability test class
